Send magic link emails over SMTP when running locally (and update instructions in example.env)
If running locally, this is the same function that was used at the deadline.

The only reason I updated emails to be sent over HTTP a day later is that sending SMTP requests from a serverless app can be pretty unreliable (even after wrapping them in Promises), and I wanted users to be able to make accounts and contribute on the production site rather than just scroll.